
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{DIFshiny}
\alias{DIFshiny}
\title{Launches the shiny app to perform DIF Analysis.}
\usage{
DIFshiny()
}
\value{No return value, launches application}
\description{
By using the functions in this package without any arguments, you can perform your Differential Item Functioning (DIF) analysis with user-friendly shiny interfaces.

}
\examples{
 if(interactive()){

 DIFshiny()

 }
}


